+++ client/public/manifest.json
@@ -2,7 +2,6 @@
   "name": "Boost Restaurant",
   "version": "1.1",
   "manifest_version": 2,
-  "key": "",
   "background": {
     "scripts": ["background.js"]
   },
